feat: database models
Added database models for customers, accounts and transactions.
Added constant values for customers and transactions.
Added a helper function that throws an error when an environment variable is not defined.
Added the logs folder as one of the excluded folders in .gitignore.
Added mongodb connection url retrieval in configuration file for environment variables.